feat: refatorar o componente EntitiesBrowser, adicionar filtros e implementar modal de confirmação de exclusão
